Rename middleware.ts to proxy.ts, move tooling configs into config/
Next 16 renamed the middleware.ts file convention to proxy.ts (the
build was already warning about this since Phase 2a). Ran the
equivalent of the official codemod by hand: file and test file
renamed, exported middleware() -> proxy(), doc references to
middleware.ts/middleware() updated to proxy.ts/proxy() (general
"Edge Middleware" mechanism references left as-is, since Next's own
docs still call the underlying feature that).

Moved jest.config.ts, jest.integration.config.ts, playwright.config.ts,
and eslint.config.mjs into frontend/config/, updating package.json's
scripts to point at the new paths and fixing each config's internal
relative paths (jest's rootDir, playwright's testDir/globalSetup/
globalTeardown, an import in playwright.config.ts). next.config.mjs
and tsconfig.json stay at the frontend/ root -- Next's CLI has no
--config flag to relocate next.config, and tsconfig.json is expected
at root by Next itself (which auto-edits it), by the TS-ESLint
parser, and by editor auto-detection.

Also deleted jsconfig.json: dead config left over from before the
project was fully TypeScript. tsconfig.json always wins when both
exist, and grepping confirmed zero `@/`-alias imports anywhere
actually depend on it.

Verified: yarn build, tsc --noEmit, yarn lint (0 errors), test:unit
(57/57), test:integration (4/4), full test:e2e (68/68), all under
Node 24.18.0, after each individual step (rename, then each config
move) to isolate which change a break would belong to.
