+++ docs/handoff/technical-decisions.md
@@ -52,7 +52,7 @@ This document answers "why did we build it this way?" for every significant tech
 **Notes:**
 - The `calendar.events` OAuth scope is sensitive, which puts the app's Google Cloud OAuth consent screen in "unverified" status (100-test-user cap, warning banner, manual test-user allowlisting) unless the consent screen's User Type is switched to **Internal**.
 
-**Token refresh is persisted in middleware, not in route handlers:** `services/auth.ts`'s `getAuth()` calls `getServerSession(authOptions)` with a single argument — next-auth v4's "RSC" code path, which uses a stub response object whose `setCookie()` is a no-op. So a route handler refreshing the Google access token via that path can never actually write the refreshed value back to the session cookie; every subsequent request would see the same stale, soon-to-expire token and refresh again, forever. `frontend/middleware.ts` owns the fix instead: it decodes the session JWT via `getToken()`, and if the access token is within 60 seconds of expiring, calls `services/googleTokenRefresh.ts`'s `refreshGoogleAccessToken()` (extracted out of `authConfig.ts`'s `jwt` callback so it has no Prisma dependency and can run on Edge middleware) and re-encodes/persists the result onto the response cookie via `next-auth/jwt`'s `encode()`. `middleware.ts`'s matcher now covers all of `/api/*` except `/api/auth/*` (NextAuth's own handler, which manages its own cookies correctly through its real route). `authConfig.ts`'s `jwt` callback still calls the same refresh helper as a fallback for the one request that actually crosses the expiry threshold — that request's own downstream refresh is redundant but harmless, since middleware has already persisted a fresh cookie for every request after it.
+**Token refresh is persisted in the proxy, not in route handlers:** `services/auth.ts`'s `getAuth()` calls `getServerSession(authOptions)` with a single argument — next-auth v4's "RSC" code path, which uses a stub response object whose `setCookie()` is a no-op. So a route handler refreshing the Google access token via that path can never actually write the refreshed value back to the session cookie; every subsequent request would see the same stale, soon-to-expire token and refresh again, forever. `frontend/proxy.ts` owns the fix instead (renamed from `middleware.ts` per Next 16's `proxy` file convention — same Edge Middleware mechanism, new name): it decodes the session JWT via `getToken()`, and if the access token is within 60 seconds of expiring, calls `services/googleTokenRefresh.ts`'s `refreshGoogleAccessToken()` (extracted out of `authConfig.ts`'s `jwt` callback so it has no Prisma dependency and can run on Edge middleware) and re-encodes/persists the result onto the response cookie via `next-auth/jwt`'s `encode()`. `proxy.ts`'s matcher now covers all of `/api/*` except `/api/auth/*` (NextAuth's own handler, which manages its own cookies correctly through its real route). `authConfig.ts`'s `jwt` callback still calls the same refresh helper as a fallback for the one request that actually crosses the expiry threshold — that request's own downstream refresh is redundant but harmless, since the proxy has already persisted a fresh cookie for every request after it.
 
 ---
 
@@ -119,7 +119,7 @@ Diagnostics (`GET /api/admin/diagnostics`, surfaced on `/admin`) checks each roo
 **Provisional tests:** features referenced in the manual QA script but not yet built (conflict detection, XLSX import, the suspend workflow's UI) get tests that lock in their *current* stub behavior rather than being skipped, tagged `@provisional-<ticket>` with a comment pointing at the exact stub line. The goal is that whoever ships the real feature finds the test immediately instead of it silently asserting the old absence of behavior forever.
 
 **Trade-offs:**
-- CI only runs Chromium (`projects: [{ name: "chromium" }]` in `playwright.config.ts`), and no automated tier touches real Zoom/Google credentials. Cross-browser rendering and live-credential behavior (a real Zoom meeting actually getting created, a real Google Calendar event actually appearing) are covered instead by a trimmed manual checklist (`docs/testing/manual-test-script-template.md`), not automation.
+- CI only runs Chromium (`projects: [{ name: "chromium" }]` in `config/playwright.config.ts`), and no automated tier touches real Zoom/Google credentials. Cross-browser rendering and live-credential behavior (a real Zoom meeting actually getting created, a real Google Calendar event actually appearing) are covered instead by a trimmed manual checklist (`docs/testing/manual-test-script-template.md`), not automation.
 - `workers: 1` — the whole E2E run shares one in-memory Mongo replica set serially rather than one per worker. Fine at this suite's size; documented as a future step if parallelism is ever needed.
 
 ---
